refactor unsorted views into separate modules
