+++ modernomad/core/urls/bookings.py
@@ -1,7 +1,7 @@
 from django.urls import re_path
 
+from modernomad.core.views.billing import submit_payment
 from modernomad.core.views.booking import *
-from modernomad.core.views.unsorted import submit_payment
 
 # urls starting in /booking get sent here.
 urlpatterns = [
