+++ modernomad/core/urls/manage.py
@@ -1,96 +1,116 @@
 from django.urls import re_path
 
-from modernomad.core.views.unsorted import *
+from modernomad.core.views import billing, booking_management, occupancy
 
 # custom management patterns
 urlpatterns = [
-    re_path(r"^payments/$", payments_today, name="location_payments_today"),
+    re_path(r"^payments/$", billing.payments_today, name="location_payments_today"),
     re_path(
-        r"^payments/(?P<year>\d+)/(?P<month>\d+)$", payments, name="location_payments"
+        r"^payments/(?P<year>\d+)/(?P<month>\d+)$",
+        billing.payments,
+        name="location_payments",
     ),
-    re_path(r"^today/$", manage_today, name="manage_today"),
-    re_path(r"bookings/$", BookingManageList, name="booking_manage_list"),
-    re_path(r"booking/create/$", BookingManageCreate, name="booking_manage_create"),
-    re_path(r"booking/(?P<booking_id>\d+)/$", BookingManage, name="booking_manage"),
+    re_path(r"^today/$", occupancy.manage_today, name="manage_today"),
+    re_path(
+        r"bookings/$", booking_management.BookingManageList, name="booking_manage_list"
+    ),
+    re_path(
+        r"booking/create/$",
+        booking_management.BookingManageCreate,
+        name="booking_manage_create",
+    ),
+    re_path(
+        r"booking/(?P<booking_id>\d+)/$",
+        booking_management.BookingManage,
+        name="booking_manage",
+    ),
     re_path(
         r"booking/(?P<booking_id>\d+)/action/$",
-        BookingManageAction,
+        booking_management.BookingManageAction,
         name="booking_manage_action",
     ),
     re_path(
         r"booking/(?P<booking_id>\d+)/paywithdrft/$",
-        BookingManagePayWithDrft,
+        booking_management.BookingManagePayWithDrft,
         name="booking_manage_pay_drft",
     ),
     re_path(
         r"booking/(?P<booking_id>\d+)/togglecomp/$",
-        BookingToggleComp,
+        booking_management.BookingToggleComp,
         name="booking_toggle_comp",
     ),
     re_path(
         r"booking/(?P<booking_id>\d+)/sendreceipt/$",
-        BookingSendReceipt,
+        booking_management.BookingSendReceipt,
         name="booking_send_receipt",
     ),
     re_path(
         r"booking/(?P<booking_id>\d+)/sendwelcome/$",
-        BookingSendWelcomeEmail,
+        booking_management.BookingSendWelcomeEmail,
         name="booking_send_welcome",
     ),
     re_path(
         r"booking/(?P<booking_id>\d+)/sendmail/$",
-        BookingSendMail,
+        booking_management.BookingSendMail,
         name="booking_send_mail",
     ),
-    re_path(r"bill/(?P<bill_id>\d+)/charge/$", BillCharge, name="bill_charge"),
-    re_path(r"bill/(?P<bill_id>\d+)/payment/$", ManagePayment, name="manage_payment"),
+    re_path(r"bill/(?P<bill_id>\d+)/charge/$", billing.BillCharge, name="bill_charge"),
     re_path(
-        r"bill/(?P<bill_id>\d+)/addbillitem/$", AddBillLineItem, name="add_bill_item"
+        r"bill/(?P<bill_id>\d+)/payment/$", billing.ManagePayment, name="manage_payment"
+    ),
+    re_path(
+        r"bill/(?P<bill_id>\d+)/addbillitem/$",
+        billing.AddBillLineItem,
+        name="add_bill_item",
     ),
     re_path(
         r"bill/(?P<bill_id>\d+)/deletebillitem/$",
-        DeleteBillLineItem,
+        billing.DeleteBillLineItem,
         name="delete_bill_item",
     ),
     re_path(
-        r"bill/(?P<bill_id>\d+)/recalculate/$", RecalculateBill, name="recalculate_bill"
+        r"bill/(?P<bill_id>\d+)/recalculate/$",
+        billing.RecalculateBill,
+        name="recalculate_bill",
     ),
     re_path(
         r"booking/(?P<booking_id>\d+)/edit/$",
-        BookingManageEdit,
+        booking_management.BookingManageEdit,
         name="booking_manage_edit",
     ),
     re_path(
         r"^subscription/(?P<subscription_id>\d+)/bill/(?P<bill_id>\d+)/sendreceipt/$",
-        SubscriptionSendReceipt,
+        billing.SubscriptionSendReceipt,
         name="subscription_send_receipt",
     ),
     re_path(
         r"^subscriptions/(?P<subscription_id>\d+)/sendmail/$",
-        SubscriptionSendMail,
+        billing.SubscriptionSendMail,
         name="subscription_send_mail",
     ),
     re_path(
         r"^subscriptions/create$",
-        SubscriptionManageCreate,
+        billing.SubscriptionManageCreate,
         name="subscription_manage_create",
     ),
     re_path(
         r"^subscriptions/(?P<subscription_id>\d+)/$",
-        SubscriptionManageDetail,
+        billing.SubscriptionManageDetail,
         name="subscription_manage_detail",
     ),
     re_path(
         r"^subscriptions/(?P<subscription_id>\d+)/update_end_date/$",
-        SubscriptionManageUpdateEndDate,
+        billing.SubscriptionManageUpdateEndDate,
         name="subscription_manage_update_end_date",
     ),
     re_path(
         r"^subscriptions/(?P<subscription_id>\d+)/generateallbills/$",
-        SubscriptionManageGenerateAllBills,
+        billing.SubscriptionManageGenerateAllBills,
         name="subscription_manage_all_bills",
     ),
     re_path(
-        r"^subscriptions/$", SubscriptionsManageList, name="subscriptions_manage_list"
+        r"^subscriptions/$",
+        billing.SubscriptionsManageList,
+        name="subscriptions_manage_list",
     ),
 ]
