+++ modernomad/core/util.py
@@ -3,7 +3,7 @@ import logging
 from django.contrib.auth.models import User
 
 from modernomad.core.models import *
-from modernomad.core.views.unsorted import monthly_occupant_report
+from modernomad.core.views.occupancy import monthly_occupant_report
 
 logger = logging.getLogger(__name__)
