+++ modernomad/urls/user.py
@@ -2,14 +2,16 @@ from django.contrib.auth import views as auth_views
 from django.urls import re_path
 
 import gather.views
-from modernomad.core.views.booking import UserBookings
-from modernomad.core.views.redirects import old_user_bookings_redirect
-from modernomad.core.views.unsorted import (
-    ListUsers,
+from modernomad.core.views.billing import (
     PeopleDaterangeQuery,
     UserAddCard,
-    UserAvatar,
     UserDeleteCard,
+)
+from modernomad.core.views.booking import UserBookings
+from modernomad.core.views.redirects import old_user_bookings_redirect
+from modernomad.core.views.user import (
+    ListUsers,
+    UserAvatar,
     UserDetail,
     UserEdit,
     email_available,
